misc/HtmlTestSuite/results/test-0438.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="https://www.w3.org/TR/css-overflow-3/#scrollable" rel="help"/>
<link href="https://github.com/w3c/csswg-drafts/issues/6230" rel="help"/>
<style>
  main table {
    display: inline-table;
    border-collapse: collapse;
    background:gray;
    font-size: 40px;
  }
  main thead {
    background: yellow;
  }
  main tbody {
    background: green;
    border: 80px dotted blue;
  }
</style>
</head><body><main>
<div id="container" style="overflow-x:auto;max-width:500px">
  <table>
    <thead>
      <tr><td>header</td>
    </tr>
    </thead><tbody>
      <tr><td>body body body body body body body body body body body</td>
    </tr>
  </tbody></table>
</div>

<script>

test(_ =&gt; {
  let container = document.querySelector("#container");
  let table = container.querySelector("table")
  assert_equals(container.offsetWidth, 500, "The container grows to the max-width.");
  assert_equals(table.offsetWidth, 500, "The table fills the container.");
  assert_equals(container.scrollWidth, 500, "The collapsed border does not exceed the scroll container.");
});
</script>
</main></body></html>
